pyproject.toml
[tool.poetry]
name = "metadata-scraper"
version = "0.1.0"
description = "Scrapes metadata about authors and affiliated institutions from given DOIs"
authors = ["Nicole T. <nbtrieu@uci.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
crossrefapi = "^1.6.0"
fastapi = "^0.109.0"
bardapi = "^0.1.39"
pubmed-author-affiliation = "^0.3"
lxml = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
